requirements.txt
Babel
chardet
decorator
docutils
ebaysdk
freezegun
gevent
greenlet
idna
Jinja2
libsass
lxml
MarkupSafe
num2words
ofxparse
passlib
Pillow
polib
psutil
psycopg2-binary
pydot
pyOpenSSL
pyparsing
PyPDF2
pyserial
python-dateutil
pytz
pyusb
qrcode
reportlab
requests
urllib3
vobject
Werkzeug
xlrd
XlsxWriter
xlwt
zeep
black
isort
qrcode[pil]
pylint
pylint-odoo
